feat(storybook)!: move catalog-picker / interactive-list-item to better homes; rename ChecklistItem → Checklist
Storybook taxonomy reshuffle landing alongside ADR-006 table top-up:

- catalog-picker: Components/Form → Components/Addables (catalog-driven
  entry selection matches Addables intent better than Form)
- InteractiveListItem: Components/Display (removed bucket) → Components/List
- ChecklistItem renamed to Checklist + moved Components/Form → Components/List.
  Component class, file names, CSS classes (bds-checklist-item → bds-checklist),
  and the export in components/ui/index.ts all updated. Doc/comment refs in
  CompletionToggle, BoardCard, and CompletionToggle.css/mdx updated to point
  at the new name.
- CompletionToggle stays in Components/Form as the atomic primitive
  (mirrors how Checkbox stays in Form alongside its labeled compositions).

ADR-006's Components subcategory table updated to reflect new homes plus
an Amendments section documenting the changes and rationale.

BREAKING CHANGE: Consumers importing ChecklistItem from @brikdesigns/bds
must rename to Checklist. Confirmed renew-pms (ViewTaskSheet.tsx) is the
only consumer affected; will be updated in a follow-up renew-pms PR after
the next BDS publish.

Pure taxonomy + rename — no behavior, prop, or styling changes. Typecheck
passes. Visually verified the four moves in Storybook.
