+++ components/ui/Checklist/index.ts
@@ -1,0 +1,2 @@
+export { Checklist, type ChecklistProps } from './Checklist';
+export { default } from './Checklist';
