--- components/ui/ChecklistItem/index.ts
@@ -1,2 +1,0 @@
-export { ChecklistItem, type ChecklistItemProps } from './ChecklistItem';
-export { default } from './ChecklistItem';
